third_party/blink/web_tests/http/tests/devtools/sources/debugger/script-failed-to-parse-expected.txt

Tests that ParsedScriptSource event is raised after compile script with syntax error.


Running: testScriptParsedEvent
Event with script-failed-to-parse.js received
